removed the mocking that are not needed, and removed video stream related code in CallingSDKWrapper mocks
